requirement.txt
SQLAlchemy
fastapi
uvicorn[standard]
python-dotenv
